Fix regex for source
Current regex will match any of the ten line after the official helper files starting with the string "source". 
Thus, in my specific case for which I declare a variable named source_filename="{...}", the linter will display a warning for it: `    ! Please avoid sourcing additional helpers after the official helpers (in this case file _filename=$(basename$upstream__url)) `.
As it should be only interested in matching the command `source` (`source path/to/file`), a space should be appended to the current regex.